move trpc action handler to a reusable fn
